+++ bin/commands/things-command.js
@@ -15,6 +15,14 @@ var _imageHelper = _interopRequireDefault(require("../image-helper"));
 
 var _theme = _interopRequireDefault(require("../theme"));
 
+var _thingsAddTrialOperation = _interopRequireDefault(require("../operations/things-add-trial-operation"));
+
+var _thingsGetDeviceOperation = _interopRequireDefault(require("../operations/things-get-device-operation"));
+
+var _thingsGetDevicesOperation = _interopRequireDefault(require("../operations/things-get-devices-operation"));
+
+var _thingsGetProductOperation = _interopRequireDefault(require("../operations/things-get-product-operation"));
+
 var _thingsListTrialOperation = _interopRequireDefault(require("../operations/things-list-trial-operation"));
 
 var _thingsRemoveTrialOperation = _interopRequireDefault(require("../operations/things-remove-trial-operation"));
@@ -75,12 +83,28 @@ class ThingsCommand extends _command.default {
             console.log(commandLineUsage(_thingsListTrialOperation.default.usage));
             break;
 
+          case 'add:trial':
+            console.log(commandLineUsage(_thingsAddTrialOperation.default.usage));
+            break;
+
           case 'remove:trial':
             console.log(commandLineUsage(_thingsRemoveTrialOperation.default.usage));
             break;
 
+          case 'get:device':
+            console.log(commandLineUsage(_thingsGetDeviceOperation.default.usage));
+            break;
+
+          case 'get:devices':
+            console.log(commandLineUsage(_thingsGetDevicesOperation.default.usage));
+            break;
+
+          case 'get:product':
+            console.log(commandLineUsage(_thingsGetProductOperation.default.usage));
+            break;
+
           default:
-            console.log(commandLineUsage([..._thingsListTrialOperation.default.usage, ..._thingsRemoveTrialOperation.default.usage]));
+            console.log(commandLineUsage([..._thingsListTrialOperation.default.usage, ..._thingsAddTrialOperation.default.usage, ..._thingsRemoveTrialOperation.default.usage, ..._thingsGetDeviceOperation.default.usage, ..._thingsGetDevicesOperation.default.usage, ..._thingsGetProductOperation.default.usage]));
         }
 
         process.exit(0);
@@ -96,8 +120,16 @@ class ThingsCommand extends _command.default {
 
       if (operation === 'list:trial') {
         await _thingsListTrialOperation.default.run();
+      } else if (operation === 'add:trial') {
+        await _thingsAddTrialOperation.default.run();
       } else if (operation === 'remove:trial') {
         await _thingsRemoveTrialOperation.default.run();
+      } else if (operation === 'get:device') {
+        await _thingsGetDeviceOperation.default.run();
+      } else if (operation === 'get:devices') {
+        await _thingsGetDevicesOperation.default.run();
+      } else if (operation === 'get:product') {
+        await _thingsGetProductOperation.default.run();
       } else {
         await _imageHelper.default.draw('chick-helps');
         console.log(`Unknown operation: ${(operation || 'undefined').code}`.warn);